metro_display/requirements.txt
pillow
certifi
